requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face Transformers (Query Assistant)
transformers>=4.30.0
accelerate>=0.26.0
sentencepiece>=0.1.99

# Web Interface
streamlit>=1.28.0
streamlit-folium>=0.15.0

# Data & ML Utilities
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Location Services & Maps
geopy>=2.3.0
requests>=2.31.0
folium>=0.14.0

# Image Processing
Pillow>=10.0.0
